# openapi.yaml
openapi: 3.0.3
info:
  title: Order Service API
  version: 1.1.0
  description: Order management service with payment integration - Choreo Deployment Tracks demo
servers:
  - url: /

paths:
  /healthz:
    get:
      summary: Health check
      operationId: healthCheck
      responses:
        "200":
          description: Service health status
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                  version:
                    type: string

  /orders:
    get:
      summary: List all orders
      operationId: listOrders
      responses:
        "200":
          description: List of orders
          content:
            application/json:
              schema:
                type: object
                properties:
                  version:
                    type: string
                  count:
                    type: integer
                  orders:
                    type: array
                    items:
                      $ref: "#/components/schemas/Order"

    post:
      summary: Create a new order
      operationId: createOrder
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateOrderRequest"
      responses:
        "201":
          description: Order created
          content:
            application/json:
              schema:
                type: object
                properties:
                  version:
                    type: string
                  order:
                    $ref: "#/components/schemas/Order"
        "400":
          description: Invalid request

  /orders/{id}:
    get:
      summary: Get order by ID
      operationId: getOrder
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Order details
          content:
            application/json:
              schema:
                type: object
                properties:
                  version:
                    type: string
                  order:
                    $ref: "#/components/schemas/Order"
        "404":
          description: Order not found

    delete:
      summary: Delete an order
      operationId: deleteOrder
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Order deleted
        "404":
          description: Order not found

  /orders/{id}/pay:
    post:
      summary: Process payment for an order
      operationId: payOrder
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PaymentRequest"
      responses:
        "200":
          description: Payment processed
          content:
            application/json:
              schema:
                type: object
                properties:
                  version:
                    type: string
                  payment:
                    $ref: "#/components/schemas/PaymentRecord"
                  order:
                    $ref: "#/components/schemas/Order"
        "404":
          description: Order not found
        "409":
          description: Order already paid

  /orders/{id}/payment:
    get:
      summary: Get payment details for an order
      operationId: getPayment
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Payment details
          content:
            application/json:
              schema:
                type: object
                properties:
                  version:
                    type: string
                  payments:
                    type: array
                    items:
                      $ref: "#/components/schemas/PaymentRecord"
        "404":
          description: No payment found

components:
  schemas:
    OrderItem:
      type: object
      required: [name, quantity, unitPrice]
      properties:
        name:
          type: string
        quantity:
          type: integer
          minimum: 1
        unitPrice:
          type: number
          format: double

    Order:
      type: object
      properties:
        id:
          type: string
        customer:
          type: string
        items:
          type: array
          items:
            $ref: "#/components/schemas/OrderItem"
        status:
          type: string
          enum: [confirmed, payment_pending, paid, shipped, cancelled]
        total:
          type: number
          format: double
        createdAt:
          type: string

    CreateOrderRequest:
      type: object
      required: [customer, items]
      properties:
        customer:
          type: string
        items:
          type: array
          items:
            $ref: "#/components/schemas/OrderItem"

    PaymentRequest:
      type: object
      required: [method]
      properties:
        method:
          type: string
          enum: [credit_card, debit_card, bank_transfer]
        cardLast4:
          type: string

    PaymentRecord:
      type: object
      properties:
        paymentId:
          type: string
        orderId:
          type: string
        amount:
          type: number
          format: double
        method:
          type: string
        cardLast4:
          type: string
        status:
          type: string
          enum: [completed, failed, refunded]
        paidAt:
          type: string