# src/main/java/com/choreo/demo/order/model/PaymentRequest.java
package com.choreo.demo.order.model;

public class PaymentRequest {

    private String method;
    private String cardLast4;

    public String getMethod() {
        return method;
    }

    public void setMethod(String method) {
        this.method = method;
    }

    public String getCardLast4() {
        return cardLast4;
    }

    public void setCardLast4(String cardLast4) {
        this.cardLast4 = cardLast4;
    }
}